Extract per-tool handlers from mcp.tools
Final PR in the mcp/tools.py decomposition sequence. Extracts all 20
per-tool handler functions to a dedicated module, no behavior change.

New file:
- tools/handlers.py (~1790 LOC) — 20 handler functions:
  _list_agreements, _search_agreements, _get_agreement,
  _get_agreement_tax_clauses, _get_agreements_summary,
  _get_agreement_trends, _get_section, _get_section_snippet,
  _get_section_snippets_batch, _get_sections_batch,
  _list_agreement_sections, _list_agreement_sections_batch,
  _search_sections, _get_section_tax_clauses,
  _suggest_clause_families, _list_filter_options,
  _get_clause_taxonomy, _get_tax_clause_taxonomy,
  _get_counsel_catalog, _get_naics_catalog.

The two server-introspection handlers (_get_server_capabilities,
_get_server_metrics) stayed in __init__.py — they reference the
capabilities-payload cluster and registry-derived state, so keeping
them adjacent avoids circular imports.

Updated one test to import _json_compatible_value from its new
home (backend.mcp.tools.shared) instead of through the package
re-export, removing the need for backwards-compat re-imports.

tools/__init__.py: 2772 -> 1024 LOC. Original tools.py was 4910 LOC,
so the 4-PR sequence (#54, #55, #56, this one) removed 79% of the
content from __init__.py into focused, single-purpose submodules.
